Added some useful features

+++ Sequence.aspx.cs
@@ -315,17 +315,52 @@ public partial class Sequence : CbstHelper
 			FileTextBox.Text = FileTextBox.Text.ToUpper();
             List<string> lsSequence = new List<string>(FileTextBox.Text.Split('\n'));
             Boolean isGroupOfTests = false;
+            Boolean isDBTypeMSSQL = false;
+            Boolean isDBTypeORACLE = false;
             for (int i = 0; i < lsSequence.Count; i++) // Remove spaces between tests and after test name
             {
-                if (lsSequence[i] == "")
+                if (lsSequence[i] == "\r")
                 {
                     lsSequence.RemoveAt(i);
+                    i--;
                     continue;
+                }
+                for (int j = 0; j < lsSequence[i].Length; j++)
+                {
+                    if (lsSequence[i][j] == ' ' && lsSequence[i][j + 1] == ' ')
+                    {
+                        lsSequence[i] = lsSequence[i].Remove(j + 1, 1);
+                        j--;
+                    }
                 }
                 while (lsSequence[i].EndsWith(" \r"))
                 {
                     lsSequence[i] = lsSequence[i].Substring(0, lsSequence[i].Length - 2);
                     lsSequence[i] += "\r";
+                }
+                if (UseLowerCaseForUsernamesAndPasswords.Checked)
+                {
+                    string[] commands = { "USER:", "PASS:" };
+                    string command;
+                    int pos;
+                    for (int j = 0; j < commands.Length; j++)
+                    {
+                        command = commands[j];
+                        pos = lsSequence[i].IndexOf(command);
+                        if (pos >= 0)
+                        {
+                            for (int k = pos + command.Length; k < lsSequence[i].Length; k++)
+                            {
+                                command = command.Insert(command.Length, lsSequence[i][k].ToString().ToLower());
+                                if (lsSequence[i][k] == '\"')
+                                {
+                                    lsSequence[i] = lsSequence[i].Remove(pos, command.Length);
+                                    lsSequence[i] = lsSequence[i].Insert(pos, command);
+                                    break;
+                                }
+                            }
+                        }
+                    }
                 }
             }
 			if (RemoveIdenticalTests.Checked)
@@ -359,12 +394,22 @@ public partial class Sequence : CbstHelper
                     if (lsSequence[i].IndexOf("{") >= 0)
                     {
                         isGroupOfTests = true;
+                        isDBTypeMSSQL = false;
+                        isDBTypeORACLE = false;
                         List<string> lsGroupOfTests = new List<string>();
                         while (isGroupOfTests)
                         {
                             i++;
                             if (lsSequence[i].IndexOf("}") >= 0)
                             {
+                                if (lsSequence[i].IndexOf("DBTYPE:MSSQL") >= 0)
+                                {
+                                    isDBTypeMSSQL = true;
+                                }
+                                else if (lsSequence[i].IndexOf("DBTYPE:ORACLE") >= 0)
+                                {
+                                    isDBTypeORACLE = true;
+                                }
                                 isGroupOfTests = false;
                                 break;
                             }
@@ -376,18 +421,28 @@ public partial class Sequence : CbstHelper
                             {
                                 int equalElements = 0;
                                 j++;
-                                while (lsSequence[j].IndexOf("}") < 0)
+                                while (true)
                                 {
                                     if (lsGroupOfTests.IndexOf(lsSequence[j]) >= 0)
                                     {
                                         equalElements++;
                                     }
                                     j++;
-                                }
-                                if (equalElements == lsGroupOfTests.Count)
-                                {
-                                    lsSequence.RemoveRange(j - equalElements - 1, equalElements + 2);
-                                    j -= equalElements + 2;
+                                    if (lsSequence[j].IndexOf("}") >= 0)
+                                    {
+                                        if ((lsSequence[j].IndexOf("DBTYPE:MSSQL") >= 0 && isDBTypeMSSQL) || (lsSequence[j].IndexOf("DBTYPE:MSSQL") < 0 && !isDBTypeMSSQL))
+                                        {
+                                            if ((lsSequence[j].IndexOf("DBTYPE:ORACLE") >= 0 && isDBTypeORACLE) || (lsSequence[j].IndexOf("DBTYPE:ORACLE") < 0 && !isDBTypeORACLE))
+                                            {
+                                                if (equalElements == lsGroupOfTests.Count)
+                                                {
+                                                    lsSequence.RemoveRange(j - equalElements - 1, equalElements + 2);
+                                                    j -= equalElements + 2;
+                                                }
+                                            }
+                                        }
+                                        break;
+                                    }
                                 }
                             }
                         }
